:sparkles: Expose version attrs on CLI
This lets you use the CLI to configure the returned values for
"swversion", "apiversion" and "datastoreversion" on the config endpoint.

These values tend to change with every Hue update even though
they often don't introduce new features or require changes to
Fargton. When you're not on the latest version the Hue app has a
nasty habit of limiting what you can do with the devices (mostly turn
things on/off). It apparently tries to force you to update. Now that
the values are configurable on the CLI we don't need to issue a new
Fargton update in lockstep with the Hue API update.